+++ packages/atlas/CHANGELOG.md
@@ -1,5 +1,9 @@
 # atlas
 
+## v0.16.0 (2021-04-07)
+
+- Added ability to supply an `assetBytes` on a `MarkerIcon`
+
 ## v0.15.0 (2021-03-19)
 
 - Added `getBounds` method to `AtlasController`
@@ -7,7 +11,7 @@
 ## v0.14.2 (2021-02-10)
 
 - Added support for `Callout` on a Map
-- Added ability to supply an `height` on a `MarkerIcon`
+- Added ability to supply a `height` on a `MarkerIcon`
 
 ## v0.14.1 (2021-02-10)